requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
distro==1.9.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
openai==1.54.3
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
together==1.3.3
tokenizers==0.20.3
tqdm==4.67.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.17.1
